+++ cookbook/sample_reset_flag.py
@@ -4,3 +4,4 @@ if __name__ == "__main__":
     persist.change_flag_extract_topic(1,0)
     persist.change_flag_affiliation_mining(1,0)
     persist.refresh()
+