+++ triplea/client/affiliation_parser_multiple.py
@@ -1,0 +1,51 @@
+from triplea.config.settings import SETTINGS
+import requests
+from triplea.service.click_logger import logger
+from urllib.parse import quote, urljoin
+
+session = requests.Session()
+
+
+def parse_affiliation_multiple(text: str, method_num: int) -> list:
+    base_url = SETTINGS.AAA_AFFILIATION_PARSER_ENDPOINT.rstrip('/') + '/'
+    query = f"?text={quote(text)}&model={method_num}"
+    url = urljoin(base_url, query)
+
+    headers = {
+        "User-Agent": SETTINGS.AAA_CLIENT_AGENT,
+    }
+
+    # To use HTTP Basic Auth with your proxy,
+    #  use the http://user:password@host.com/ syntax:
+    if SETTINGS.AAA_PROXY_HTTP is not None:
+        proxy_servers = {
+            "http": SETTINGS.AAA_PROXY_HTTP,
+            "https": SETTINGS.AAA_PROXY_HTTPS,
+        }
+    else:
+        proxy_servers = None
+
+    # sending get request and saving the response as response object
+    try:
+        r = session.get(url=url, headers=headers, proxies=proxy_servers)
+
+    except Exception:
+        raise Exception("Connection Error.")
+
+    # extracting data in json format
+    try:
+        if r.status_code == 200:
+            data = r.json()
+            if "status" in data:
+                return data["result"]
+            else:
+                logger.ERROR("status not exist.")
+                raise
+        else:
+            logger.ERROR(f"ERROR : {r.status_code}")
+            logger.ERROR(f"Reason : {r.reason}")
+
+    except Exception as ex:
+        logger.ERROR(f"Error : {ex}")
+        logger.ERROR(f"{type(r)}  {r} ")
+        raise